refactor: Move out components from App
